Remove bogus test
