+++ src/middlewared/middlewared/test/integration/assets/privilege.py
@@ -15,3 +15,23 @@ def privilege(data):
             call("privilege.delete", privilege["id"])
         except InstanceNotFound:
             pass
+
+
+
+@contextlib.contextmanager
+def raw_privilege(data):
+    """Insert a privilege row directly into the datastore, ignoring the validation.
+    """
+    id_ = call("datastore.insert", "account.privilege", {
+        "builtin_name": None,
+        "name": "Test raw",
+        "local_groups": [],
+        "ds_groups": [],
+        "roles": [],
+        "web_shell": False,
+        **data,
+    })
+    try:
+        yield id_
+    finally:
+        call("datastore.delete", "account.privilege", id_)
